% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_for_inflation.R
\name{adjust_for_inflation}
\alias{adjust_for_inflation}
\title{Adjust for Inflation}
\usage{
adjust_for_inflation(value_to_adjust, from_cpi, to_cpi)
}
\arguments{
\item{value_to_adjust}{The amount to adjust.}

\item{from_cpi}{The starting counsumer price index value.}

\item{to_cpi}{The ending consumer price index value.}
}
\value{
A numeric.
}
\description{
Adjust value for inflation, using to different consumer price index values.
}
\examples{
adjust_for_inflation(100, 50, 100)
}
