% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm_options.R
\name{arm_opts}
\alias{arm_opts}
\title{Setting Options on ARM loan}
\usage{
arm_opts(fixed_term = "fixed_term", period = "period",
  lifetime_cap = "lifetime_cap", lifetime_floor = "lifetime_floor",
  periodic_cap = "periodic_cap", periodic_floor = "periodic_floor",
  margin = "margin", index = c(...))
}
