% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_standard_rate.R
\name{calculate_standard_rate}
\alias{calculate_standard_rate}
\title{Calculate the Standard Rate of Capital Income}
\usage{
calculate_standard_rate(government_borrowing_rate_table = mortgage::gbr_table)
}
\arguments{
\item{government_borrowing_rate_table}{A table of the government borrowing
rate, from the function \code{get_governemnt_borroing_rate()}.}
}
\value{
A tibble.
}
\description{
Calculate the standard rate of capital income, using \href{https://www.avanza.se/lar-dig-mer/avanza-akademin/skatt-deklaration/hur-beskattas-en-kapitalforsakring.html}{the method described by Avanza}.
}
\examples{
\dontrun{
calculate_standard_rate()
}
}
\seealso{
\code{\link[=get_government_borrowing_rate]{get_government_borrowing_rate()}}

\link{gbr_table}
}
