% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpi_table}
\alias{cpi_table}
\title{A Table of the Swedish Consumer Price Index Between 1980-01-01 and 2024-09-01}
\format{
\subsection{\code{cpi_table} A data frame with 537 rows and 2 columns:}{

\describe{
\item{date}{Date of observation.}
\item{cpi}{Consumer price index.}
}
}
}
\source{
\url{https://www.scb.se/hitta-statistik/statistik-efter-amne/priser-och-konsumtion/konsumentprisindex/konsumentprisindex-kpi/}
}
\usage{
cpi_table
}
\description{
A table of \href{https://www.scb.se/hitta-statistik/statistik-efter-amne/priser-och-konsumtion/konsumentprisindex/konsumentprisindex-kpi/}{the Swedish consumer price index}.
}
\keyword{datasets}
