% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_interest_rate.R
\name{daily_interest_rate}
\alias{daily_interest_rate}
\title{Calculate Daily Compounding Interest Rate}
\usage{
daily_interest_rate(yearly_interest_rate, calendar_year)
}
\arguments{
\item{yearly_interest_rate}{The yearly interest rate.}

\item{calendar_year}{The calendar year.}
}
\value{
The daily interest rate.
}
\description{
Calculate daily compounding interest rate from a yearly rate and calendar
year. Accounts for leap years. Uses the following formula: \deqn{\text{daily
interest rate} = (1 + \text{yearly interest rate})^{(\frac{1}{365 + 1 \text{
if calendar year is a leap year}})} - 1}
}
\examples{
daily_rate <- daily_interest_rate(0.05, 2001)
}
