% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debt_ratio.R
\name{debt_ratio}
\alias{debt_ratio}
\title{Calculate Debt Ratio}
\usage{
debt_ratio(remaining_principal, household_gross_income)
}
\arguments{
\item{remaining_principal}{Remaining principal.}

\item{household_gross_income}{Household gross income.}
}
\value{
A quotient.
}
\description{
Calculate debt ratio as defined by
\href{https://www.fi.se/contentassets/690c9d8430bd4961b6073dedf4f9105f/fs1616-k2024.pdf}{Finansinspektionen}.
}
\examples{
debt_ratio(400, 100)
}
