% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbr_table}
\alias{gbr_table}
\title{A Table of the Swedish Government Borrowing Rate Between 1986-12-05 and
2024-11-08}
\format{
\subsection{\code{gbr_table} A data frame with 1,980 rows and 3 columns:}{

\describe{
\item{date}{Date of observation.}
\item{gbr}{Government borrowing rate.}
\item{yearly_mean}{Yearly mean up to the the current observation}
}
}
}
\source{
\url{https://www.riksgalden.se/globalassets/dokument_sve/statslaneranta/slr-historisk-statslaneranta-csv.csv}
}
\usage{
gbr_table
}
\description{
A table of \href{https://www.riksgalden.se/sv/var-verksamhet/statslanerantan/}{the Swedish government borrowing rate}.
}
\keyword{datasets}
