% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpi.R
\name{get_cpi}
\alias{get_cpi}
\title{Get Consumer Price Index from Statistics Sweden}
\usage{
get_cpi()
}
\value{
A tibble of historical consumer price index values. Note that the
dates in the return tibble are set to the first day of their respective
months, due to limitations in how dates are stored. The values apply to the
entire months in question.
}
\description{
Downloads consumer price index from \href{https://www.statistikdatabasen.scb.se}{the Statistics Sweden API}.
}
\examples{
\dontrun{
cpi_data <- get_cpi()
}
}
