% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_list_rates.R
\name{get_current_list_rates}
\alias{get_current_list_rates}
\title{Get Current List Rates from Swedish Banks}
\usage{
get_current_list_rates()
}
\value{
A tibble of list rates, banks, and the date they were updated.
}
\description{
Downloads the current list rates from the Swedish banks listed on \href{https://www.konsumenternas.se/konsumentstod/jamforelser/lan--betalningar/lan/jamfor-borantor/}{this site}.
}
\examples{
\dontrun{
list_rates <- get_current_list_rates()
}
}
