% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_government_borrowing_rate.R
\name{get_government_borrowing_rate}
\alias{get_government_borrowing_rate}
\title{Get Government Borrowing Rate}
\usage{
get_government_borrowing_rate()
}
\value{
A tibble.
}
\description{
Downloads the government borrowing rate from [the Swedish National Debt
Office](https://www.riksgalden.se/sv/var-verksamhet/statslanerantan/).
}
\examples{
\dontrun{
gbr_table <- get_government_borrowing_rate()
}
}
