% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interest_markup.R
\name{get_interest_markup}
\alias{get_interest_markup}
\title{Get Daily Interest Markup}
\usage{
get_interest_markup()
}
\value{
A numeric.
}
\description{
Get mean markup on interest used by the banks Swedbank, Nordea, SEB, and
Swedbank, as compared to the Swedish Naional Bank’s policy rate, converted to
a daily interest rate.
}
\examples{
\dontrun{
current_mean_interest_markup <- get_interest_markup()
}
}
