% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_old_omx.R
\name{get_old_omx}
\alias{get_old_omx}
\title{Get OMXS30 Data from Nasdaq OMX Nordic’s Old Site}
\usage{
get_old_omx(
  instrument = "SE0000337842",
  start_date = "1986-09-30",
  sleep_time = 4
)
}
\arguments{
\item{instrument}{Instrument ID. Defaults to the OMXS30 index.}

\item{start_date}{Start date. Defaults to 1986-09-30}

\item{sleep_time}{Seconds to wait between scraping commands.}
}
\description{
Download historical data on the OMXS30 index form the Nasdaq OMX Nordic site
that closes on November 14th 2024. This data stretches back to 1986, whereas
the data on the new site only goes back ten years. Data downloaded using this
function can be merged with data from the new site. A separate function
handles data import from the new site. The current function will not be
exported with the package, but merely used as an internal support function to
download package data.
}
