% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omx.R
\name{get_omx}
\alias{get_omx}
\title{Get OMXS30 Data from Nasdaq OMX Nordic}
\usage{
get_omx(
  instrument = "SE0000337842",
  start_date = format((Sys.time() - lubridate::years(10) - lubridate::days(1)),
    "\%Y-\%m-\%d"),
  end_date = format((Sys.time() - lubridate::days(1)), "\%Y-\%m-\%d")
)
}
\arguments{
\item{instrument}{Instrument ID. Defaults to the OMXS30 index.}

\item{start_date}{Start date. Defaults to ten years and one day before today.}

\item{end_date}{End date. Defaults to one day before today.}
}
\value{
A tibble of historical OMXS30 data.
}
\description{
Download historical data on the OMXS30 index form the Nasdaq OMX Nordic site.
}
\examples{
\dontrun{
omx_data <- get_omx()
}
}
