% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_policy_rate.R
\name{get_policy_rate}
\alias{get_policy_rate}
\title{Get Historical Policy Rates from the Swedish Central Bank}
\usage{
get_policy_rate(
  series_id = "SECBREPOEFF",
  start_date = "1994-06-01",
  end_date = strftime(Sys.time(), "\%Y-\%m-\%d")
)
}
\arguments{
\item{series_id}{Series ID used by the Swedish Central Bank. Defaults to the
policy rate, i.e. SECBREPOEFF.}

\item{start_date}{Start date. Defaults to 1994-06-01.}

\item{end_date}{End date. Defaults to today.}
}
\value{
A tibble of historical policy rates.
}
\description{
Downloads the Swedish policy rate using \href{https://developer.api.riksbank.se}{the Swedish Central Banks’s API}.
}
\examples{
\dontrun{
policy_rates <- get_policy_rate()
}
}
