% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_cost.R
\name{insurance_cost}
\alias{insurance_cost}
\title{Calculate Risk Cost for Kapitalforsakring at Avanza Based on Age}
\usage{
insurance_cost(age, amount)
}
\arguments{
\item{age}{Age.}

\item{amount}{Amount in kapitalforsakring.}
}
\value{
A numeric.
}
\description{
Calculate the monthly risk cost of a [kapitalforsakring at
Avanza](https://www.avanza.se/konton-lan-prislista/konton/kapitalforsakring/fordjupad-information-kapitalforsakring.html#riskkostnad)
based on age and insured amount.
}
\examples{
cost <- insurance_cost(25, 100000)
}
