% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_first_weekday.R
\name{is_first_weekday}
\alias{is_first_weekday}
\title{Check Whether Date is First Weekday}
\usage{
is_first_weekday(date)
}
\arguments{
\item{date}{A date.}
}
\value{
A Boolean.
}
\description{
Takes a date an returns True if it is the first weekday of the month.
}
\examples{
as.Date('2014-01-03') |> is_first_weekday()
}
