% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loan.R
\name{Loan}
\alias{Loan}
\title{R6 Class Representing a Loan}
\description{
A loan used to create and calculate various mortgage statistics.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{principal}}{Principal amount of the loan}

\item{\code{interest}}{Interest rate of the loan}

\item{\code{term}}{Term of the loan}

\item{\code{term_unit}}{Unit of the term}

\item{\code{compounded}}{Compounding method}

\item{\code{currency}}{Currency symbol}

\item{\code{n_periods}}{Number of periods}

\item{\code{schedule}}{Amortization schedule}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Loan-new}{\code{Loan$new()}}
\item \href{#method-Loan-total_principal}{\code{Loan$total_principal()}}
\item \href{#method-Loan-total_interest}{\code{Loan$total_interest()}}
\item \href{#method-Loan-total_paid}{\code{Loan$total_paid()}}
\item \href{#method-Loan-interest_to_principal}{\code{Loan$interest_to_principal()}}
\item \href{#method-Loan-interest_to_paid}{\code{Loan$interest_to_paid()}}
\item \href{#method-Loan-years_to_pay}{\code{Loan$years_to_pay()}}
\item \href{#method-Loan-apr}{\code{Loan$apr()}}
\item \href{#method-Loan-apy}{\code{Loan$apy()}}
\item \href{#method-Loan-monthly_payment}{\code{Loan$monthly_payment()}}
\item \href{#method-Loan-amortize}{\code{Loan$amortize()}}
\item \href{#method-Loan-split_payment}{\code{Loan$split_payment()}}
\item \href{#method-Loan-summarize}{\code{Loan$summarize()}}
\item \href{#method-Loan-print}{\code{Loan$print()}}
\item \href{#method-Loan-clone}{\code{Loan$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Loan-new"></a>}}
\if{latex}{\out{\hypertarget{method-Loan-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Loan} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Loan$new(
  principal,
  interest,
  term,
  term_unit = "years",
  compounded = "monthly",
  currency = "$"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{principal}}{Principal amount of the loan}

\item{\code{interest}}{Interest rate of the loan}

\item{\code{term}}{Term of the loan}

\item{\code{term_unit}}{Unit of the term}

\item{\code{compounded}}{Compounding method}

\item{\code{currency}}{Currency symbol}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Loan-total_principal"></a>}}
\if{latex}{\out{\hypertarget{method-Loan-total_principal}{}}}
\subsection{Method \code{total_principal()}}{
Return the principal amount of the loan
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Loan$total_principal()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Loan-total_interest"></a>}}
\if{latex}{\out{\hypertarget{method-Loan-total_interest}{}}}
\subsection{Method \code{total_interest()}}{
Return the total interest paid on the loan
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Loan$total_interest()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Loan-total_paid"></a>}}
\if{latex}{\out{\hypertarget{method-Loan-total_paid}{}}}
\subsection{Method \code{total_paid()}}{
Return the total amount paid on the loan
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Loan$total_paid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Loan-interest_to_principal"></a>}}
\if{latex}{\out{\hypertarget{method-Loan-interest_to_principal}{}}}
\subsection{Method \code{interest_to_principal()}}{
Return the percentage of the total principal that is interest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Loan$interest_to_principal()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Loan-interest_to_paid"></a>}}
\if{latex}{\out{\hypertarget{method-Loan-interest_to_paid}{}}}
\subsection{Method \code{interest_to_paid()}}{
Return the percentage of the total amount that is interest
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Loan$interest_to_paid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Loan-years_to_pay"></a>}}
\if{latex}{\out{\hypertarget{method-Loan-years_to_pay}{}}}
\subsection{Method \code{years_to_pay()}}{
Return the number of years to pay off the loan
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Loan$years_to_pay()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Loan-apr"></a>}}
\if{latex}{\out{\hypertarget{method-Loan-apr}{}}}
\subsection{Method \code{apr()}}{
Return the annual percentage rate (APR) of the loan
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Loan$apr()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Loan-apy"></a>}}
\if{latex}{\out{\hypertarget{method-Loan-apy}{}}}
\subsection{Method \code{apy()}}{
Return the annual percentage yield (APY) of the loan
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Loan$apy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Loan-monthly_payment"></a>}}
\if{latex}{\out{\hypertarget{method-Loan-monthly_payment}{}}}
\subsection{Method \code{monthly_payment()}}{
Return the monthly payment of the loan
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Loan$monthly_payment()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Loan-amortize"></a>}}
\if{latex}{\out{\hypertarget{method-Loan-amortize}{}}}
\subsection{Method \code{amortize()}}{
Return the amortization schedule of the loan
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Loan$amortize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Loan-split_payment"></a>}}
\if{latex}{\out{\hypertarget{method-Loan-split_payment}{}}}
\subsection{Method \code{split_payment()}}{
Return the interest and principal payment for a split payment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Loan$split_payment(payment_number, amount)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{payment_number}}{The number of the payment}

\item{\code{amount}}{The amount of the payment}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Loan-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-Loan-summarize}{}}}
\subsection{Method \code{summarize()}}{
Print a summary of the loan
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Loan$summarize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Loan-print"></a>}}
\if{latex}{\out{\hypertarget{method-Loan-print}{}}}
\subsection{Method \code{print()}}{
Print the loan object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Loan$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Loan-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Loan-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Loan$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
