% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{LoanComparison}
\alias{LoanComparison}
\title{R6 Class Representing a Loan Comparison}
\description{
A class used to compare multiple loan objects.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{loans}}{List of loan objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LoanComparison-new}{\code{LoanComparison$new()}}
\item \href{#method-LoanComparison-compare_total_interest}{\code{LoanComparison$compare_total_interest()}}
\item \href{#method-LoanComparison-compare_monthly_payments}{\code{LoanComparison$compare_monthly_payments()}}
\item \href{#method-LoanComparison-compare_total_payments}{\code{LoanComparison$compare_total_payments()}}
\item \href{#method-LoanComparison-compare_total_payments_diff}{\code{LoanComparison$compare_total_payments_diff()}}
\item \href{#method-LoanComparison-summarize}{\code{LoanComparison$summarize()}}
\item \href{#method-LoanComparison-print}{\code{LoanComparison$print()}}
\item \href{#method-LoanComparison-clone}{\code{LoanComparison$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoanComparison-new"></a>}}
\if{latex}{\out{\hypertarget{method-LoanComparison-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{LoanComparison} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoanComparison$new(loans)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{loans}}{List of loan objects}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoanComparison-compare_total_interest"></a>}}
\if{latex}{\out{\hypertarget{method-LoanComparison-compare_total_interest}{}}}
\subsection{Method \code{compare_total_interest()}}{
Compare the total interest of the loans
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoanComparison$compare_total_interest()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoanComparison-compare_monthly_payments"></a>}}
\if{latex}{\out{\hypertarget{method-LoanComparison-compare_monthly_payments}{}}}
\subsection{Method \code{compare_monthly_payments()}}{
Compare the monthly payments of the loans
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoanComparison$compare_monthly_payments()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoanComparison-compare_total_payments"></a>}}
\if{latex}{\out{\hypertarget{method-LoanComparison-compare_total_payments}{}}}
\subsection{Method \code{compare_total_payments()}}{
Compare the total payments of the loans
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoanComparison$compare_total_payments()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoanComparison-compare_total_payments_diff"></a>}}
\if{latex}{\out{\hypertarget{method-LoanComparison-compare_total_payments_diff}{}}}
\subsection{Method \code{compare_total_payments_diff()}}{
Compare difference of total payments between two or more loans in matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoanComparison$compare_total_payments_diff()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoanComparison-summarize"></a>}}
\if{latex}{\out{\hypertarget{method-LoanComparison-summarize}{}}}
\subsection{Method \code{summarize()}}{
Print a summary of the loan comparison
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoanComparison$summarize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoanComparison-print"></a>}}
\if{latex}{\out{\hypertarget{method-LoanComparison-print}{}}}
\subsection{Method \code{print()}}{
Print the loan object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoanComparison$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoanComparison-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LoanComparison-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoanComparison$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
