% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{long_omxs30}
\alias{long_omxs30}
\title{OMXS30 Index Data Since 1986-09-30 and 2024-11-08}
\format{
\subsection{\code{long_omxs30} A data frame with 9,562 rows and 7 columns:}{

\describe{
\item{date}{Date of observation.}
\item{high_price}{High price.}
\item{low_price}{Low price.}
\item{closing_price}{Closing price.}
\item{average_price}{Average price.}
\item{total_volume}{Total volume.}
\item{turnover}{Turnover.}
}
}
}
\source{
\url{https://www.nasdaqomxnordic.com/index/historiska_kurser?Instrument=SE0000337842}
\url{https://www.nasdaq.com/european-market-activity/indexes/omxs30?id=SE0000337842}
}
\usage{
long_omxs30
}
\description{
\href{https://www.nasdaq.com/european-market-activity/indexes/omxs30?id=SE0000337842}{OMX Stockholm 30 Index}
prices between 1986-09-30 and 2024-11-08.
}
\keyword{datasets}
