% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvr.R
\name{lvr}
\alias{lvr}
\title{Calculate Loan to Value Ratio}
\usage{
lvr(remaining_principal, asset_value)
}
\arguments{
\item{remaining_principal}{Remaining principal.}

\item{asset_value}{Value of asset.}
}
\value{
A quotient.
}
\description{
Divides remaining principal by asset value.
}
\examples{
lvr(50, 100)
}
