% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_daily_table.R
\name{make_daily_table}
\alias{make_daily_table}
\title{Make a Table of Daily Interest Rates and Rate of Change in OMXS30 and CPI}
\usage{
make_daily_table()
}
\value{
A table of daily rate of change of interest, OMXS30, and CPI.
}
\description{
Makes a table of daily rates of change in \href{https://www.riksbank.se/sv/statistik/rantor-och-valutakurser/styrranta-in--och-utlaningsranta/}{the Swedish Central Bank policy rate},
\href{https://www.nasdaq.com/european-market-activity/indexes/omxs30?id=SE0000337842}{the OMX Stockholm 30 Index},
and \href{https://www.scb.se/hitta-statistik/statistik-efter-amne/priser-och-konsumtion/konsumentprisindex/konsumentprisindex-kpi/}{the Swedish consumer price index}.
}
\examples{
\dontrun{
daily_table <- make_daily_table()
}
}
