% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimum_monthly_payment.R
\name{minimum_monthly_payment}
\alias{minimum_monthly_payment}
\title{Calculate Minimum Required Payment Towards Principal}
\usage{
minimum_monthly_payment(initial_principal, loan_to_value_ratio, debt_ratio)
}
\arguments{
\item{initial_principal}{Initial principal amount.}

\item{loan_to_value_ratio}{Ratio of asset value to remaining principal.}

\item{debt_ratio}{Ratio of debt to household gross income.}
}
\value{
A numeric.
}
\description{
Calculate the minimum required payment towards the principal of a housing
loan, \href{https://www.fi.se/contentassets/690c9d8430bd4961b6073dedf4f9105f/fs1616-k2024.pdf}{as described by Finansinspektionen}.
}
\examples{
minimum_monthly_payment(500, 0.5, 3)
}
\seealso{
\code{\link[=lvr]{lvr()}}

\code{\link[=debt_ratio]{debt_ratio()}}
}
