% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/months_unitl.R
\name{months_until}
\alias{months_until}
\title{Calculate the Number of Months Between Two Dates}
\usage{
months_until(to_date, from_date = Sys.Date())
}
\arguments{
\item{to_date}{Second date.}

\item{from_date}{First date. Defaults to today.}
}
\value{
A numeric.
}
\description{
Calculates the number of months between two dates.
}
\examples{
number_of_months <- months_until(Sys.Date() + 365)
}
