% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_of_change.R
\name{rate_of_change}
\alias{rate_of_change}
\title{Calculate Rate of Change as a Geometric Mean}
\usage{
rate_of_change(start_value, end_value, time_delta)
}
\arguments{
\item{start_value}{Value at start of period.}

\item{end_value}{Value at end of period.}

\item{time_delta}{Length of period.}
}
\value{
The rate of change per time unit.
}
\description{
Calculate the rate of change between two values per unit of time. Uses the
following formula to calculate a geometric mean: \deqn{\text{rate of change}
= \sqrt[\Delta_{\text{time}}]{\frac{\text{end value}}{\text{start value}}} -
1}
}
\examples{
rate_of_change(100, 200, 31)
}
