% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remaining_principal.R
\name{remaining_principal}
\alias{remaining_principal}
\title{Calculate Remaining Principal after Payment}
\usage{
remaining_principal(remaining_principal, payment_towards_principal)
}
\arguments{
\item{remaining_principal}{Remaining principa before payment.}

\item{payment_towards_principal}{Payment.}
}
\value{
A difference.
}
\description{
Subtracts payment from principal before payment.
}
\examples{
remaining_principal(100, 10)
}
