% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_sum.R
\name{standard_sum}
\alias{standard_sum}
\title{Calculate the Absolute Value of Tax Based on a Standard Rate of Capital
Income}
\usage{
standard_sum(date, standard_rate_of_capital_income, transaction_total)
}
\arguments{
\item{date}{Date of transaction.}

\item{standard_rate_of_capital_income}{The standard rate of capital income.}

\item{transaction_total}{The transaction total.}
}
\value{
A numeric.
}
\description{
Calculate the absolute tax for a transaction based on a tandard rate of
capital income, using \href{https://www.avanza.se/lar-dig-mer/avanza-akademin/skatt-deklaration/hur-beskattas-en-kapitalforsakring.html}{the method described by Avanza}.
}
\examples{
standard_sum(as.Date('2024-01-01'), 0.0362, 100000)
}
