% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daily_table}
\alias{daily_table}
\title{A Table of Daily Interest Rate and Rates of Change in OMXS30 and CPI between
1994-06-01 and 2024-08-01}
\format{
\subsection{\code{daily_table} A data frame with 11,020 rows and 5 columns:}{

\describe{
\item{date}{Date of observation.}
\item{interest}{Daily Swedish policy rate.}
\item{omxs30}{Daily rate of change in the OMXS 30 index.}
\item{cpi}{Daily rate of change in the Swedish consumer price index. Note that observations are only made once per month.}
\item{standard_rate}{Standard rate of capital income.}
}
}
}
\source{
\url{https://www.riksbank.se/sv/statistik/rantor-och-valutakurser/styrranta-in--och-utlaningsranta/}
\url{https://www.nasdaq.com/european-market-activity/indexes/omxs30?id=SE0000337842}
\url{https://www.scb.se/hitta-statistik/statistik-efter-amne/priser-och-konsumtion/konsumentprisindex/konsumentprisindex-kpi/}
}
\usage{
daily_table
}
\description{
A table of daily rates of change in \href{https://www.riksbank.se/sv/statistik/rantor-och-valutakurser/styrranta-in--och-utlaningsranta/}{the Swedish Central Bank policy rate},
\href{https://www.nasdaq.com/european-market-activity/indexes/omxs30?id=SE0000337842}{the OMX Stockholm 30 Index},
\href{https://www.scb.se/hitta-statistik/statistik-efter-amne/priser-och-konsumtion/konsumentprisindex/konsumentprisindex-kpi/}{the Swedish consumer price index},
and \href{https://www.avanza.se/lar-dig-mer/avanza-akademin/skatt-deklaration/hur-beskattas-en-kapitalforsakring.html}{the standard rate of capital income},
calculated from \href{https://www.riksgalden.se/sv/var-verksamhet/statslanerantan/}{the government borrowing rate}.
}
\keyword{datasets}
