% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpi.R
\name{get_cpi}
\alias{get_cpi}
\title{Get Consumer Price Index from Statistics Sweden}
\usage{
get_cpi()
}
\value{
A tibble of historical consumer price indes values.
}
\description{
Downloads consumer price index from \href{https://www.statistikdatabasen.scb.se}{the Statistics Sweden API}.
}
\examples{
\dontrun{
cpi_data <- get_cpi()
}
}
